target zone = optimized CLEAR zone
The tooltip already said "optimize for clearing" but it should beat the user over the head with that a bit more. And well also, the save-load function should pull the last cleared zone rather than the portal zone to match the tooltip and the reasoning for having this input in the first place ;p

I didn't worry about +/- 1 too much before, but it's critical to get this exactly right for Observation. 148 vs 149 is a massive difference, for example. (since clearing 149 gets you the free trinkets for 150)