don't hide non-maxed tributes
Even if Greed is maxed we should NOT hide tribute count if <1250, as this would result in an invisible setting that drastically over-weights resource boosts by giving them credit for boosting greed via more tributes.

Discovered this bug because I switched from a battle spec to a radon spec and got 4 MORE levels of Motivation, the unchecked the hide maxed button and found I'd locked 1200 tributes. Setting that field to 1250 knocked off 5 levels of Mot. YIKE.