prettify radon input fields
Make Rn per run and Rn for perks fields "text" type instead of "number" and do my own legalization, god have mercy on my soul.

Replace direct call to evaluatePerks from event handlers with parseAndEvaluate, a new function that queries the event that called it and (if it was called from a 'text' type field) stuffs the entered value into the new rawValue field. This field is also set for 'text' fields when loading a save with the field unlocked, when loading the localstorage on Surky page-load, and at no other time ever ever EVER. 'text' fields always use the rawValue for the calculation. I hope and pray to $deity that this rawValue is always filled in and correct when needed :notlikeblob:

'text' type aux fields are prettified for display by simply calling toExponential(4) on the raw value.