trinket value methodology fix
Previous trinket value calculation used the raw gain of the next level as the gain per level, which is fine under the assumption the gain curve is locally flat and we're evaluating small trinket gains vs. small perk level gains on the order of 1 level or less. But it breaks BADLY (massive proliferation of NaNs) if mot/pow/tough are significantly over- or under-leveled compared to the assumption that they are co-optimized at their current levels and "normal" valuations, which can occur if the user locks a higher or lower level than normal (weird but possible) or more to the point in the Trappa setting where motivation is massively devalued.

Turns out this lazy approximation was never necessary in the first place, as calculating the actual levels needed to achieve a given gain is just a couple lines of simple algebra :facepalm: - so, now I've done that, and trinket valuation will account for the user's implied valuation of these perks even if they're badly over- or under-leveled due to locking or a setting like Trappa.

Also this allows devaluing motivation even further in trappa.