tauntimp correction to housing gains
Prior assumption: More resources give more housing proportional to additional housing built at target zone. This isn't true with Tauntimps. Earlier housing is better. It's effectively as if our base housing for this purpose is smaller. Working through the math, to a good approximation (for large target zones) it's just a fixed offset based on the target zone. Here we calculate that fixed offset and use it to provide a better estimate of housing gains from additional resources.