Rn/run now based on exponential moving average of recent farming runs
Rn/run is a critically important stat for perk valuation after Obs is unlocked, but it's very easy to import a save with a garbage value in game.resources.radon.owned and fuck it up. So this change implements an exponentially weighted moving average over the last 10 farming runs imported.

A "farming run" is considered anything over 5% of the currently entered Rn/run value. Doubtful (at least in the paradigm where dailies give negligible Rn) that this would ever exclude a real farming run, and 5% is chosen as an extremely conservative guess of what a non-farming run would never exceed post-Obs (with no Rn challenge boosts).

Additionally, if the entered Rn/run is greater than lifetime Rn we assume the user doesn't want this to be their permanent Rn/run value, and count the current save as a "farming" run so long as it got at least a millionth of lifetime Rn. This will promptly overwrite the erroneous input (unless it's locked).