don't autofill perk string after autobuy
At least one user strongly prefers not to have their clipboard overwritten on autobuy, which seems reasonable. So we no longer fill the perk string on autobuy, only on export.

If there's separate appetite for autobuy filling and copying the perk string, I could add a setting later.

Meanwhile there's a secret autoFill setting that fills the perk string on autobuy but does not copy it.