frenzy heuristic for ~deathless vs. 5-hit deaths
It's been pointed out that trinkets scale much more poorly than S3, and that this means it's less worthwhile to push slow zones when trinkets are the majority of farming value, and thus GB is not always used in the afterpush.

We use the heuristic here, that if there is NO clear weight (i.e. we're purely farming) and we're not asking for a pure resource preset (where we're clearly not using GB regardless), we should only use 5-hit deaths for frenzy if a 3-minute zone is worth running (based on the S3 and trinket gains from this zone), using the gross approximation of 1 minute per zone to the target zone.